requirements.txt
altgraph==0.17.4
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
importlib-metadata==7.0.1
packaging==23.2
pefile==2023.2.7
py-mon==2.0.3
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.31.0
urllib3==2.1.0
watchdog==3.0.0
zipp==3.17.0
